requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
spacy
networkx
nltk
rouge-score
bert-score
datasets
streamlit